[ai-registry-adaptor] Update unit tests for ai-registry rename
